jitsi_middlelayer/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
faker==18.13.0
